doctorcontroller

+++ doctorsappoinment/src/main/java/com/example/demo/repo/DoctorRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Doctor;
 
-public interface DoctorRepo  extends JpaRepository<Doctor, Long>{
+public interface DoctorRepo  extends JpaRepository<Doctor, Integer>{
 
 }
