+++ doctorsappoinment/src/main/java/com/example/demo/repo/PatientRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.patient;
 
-public interface PatientRepo extends JpaRepository<patient, Long> {
+public interface PatientRepo extends JpaRepository<patient, Integer> {
 
 }
